tools: build_apps: fix make warning due to unused TEST_GROUPS variable
TEST_GROUPS can be included in the unit-test-app config file, to
restrict the list of tests to be executed. However this option is not
used at build time, so adding it as a CMake variable along with
TEST_COMPONENTS and other options results in a CMake warning.
Fix by not passing this variable to CMake, and not including it in the
final sdkconfig file.