ci: iperf: recognize report results from newer iperf versions
The iperf version used in older version of CI docker images had the
following output format:
[  3]  0.0-10.0 sec  1.25 MBytes  1.05 Mbits/sec   0.000 ms    0/  892 (0%)
The newer iperf version which was recently included in the container
prints more digits after the decimal point:
[  3] 0.0000-10.0148 sec  1.25 MBytes  1.05 Mbits/sec   0.002 ms    0/  895 (0%)
The regular expression to match this line expected a single zero after
the decimal point, so the new format no longer matches.
The fix is to expect any number of digits in the fractional part.