examples: sd_card: fix example test not getting the expected name
SD card example prints the line which looks like:

   Name: 00000

If only part of the line is received, the regular expression in the
test still matches, and partial name is obtained. Later the example
test expects to see the same name in the file read from the SD card.
This check fails, because only partial name was received earlier.

Fix by adding a CR to the regular expression so that the entire name
is matched.